requirements.txt
# Core
numpy==1.26.4
opencv-python==4.11.0.86
python-dotenv==1.1.1
google-generativeai

# Streamlit viewer
streamlit==1.33.0
altair==5.1.2
pandas==2.0.3
pyarrow==14.0.2
